(#18268) (#21243) (#22368) (#22700) windows installer improvements
Adding minimum version information to the documentation 
relating to new and unreleased versions of puppet code.

#22700 adds documentation relating to adding Puppet 
to the path variable automatically and documents the 
first version of the installer that allows that.

#21243 - documentation surrounding the puppet agent 
account user and how it is specified in the installer.

#22368 - enhances the previously submitted documentation 
on 22368 with the puppet agent variable to add the 
minimum version.

#18628 - adds the documentation surrounding the 
startup mode for windows and the different items you
can pass to the installer.